+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - Add support for the following special CQL commands in `CassandraStatement`:
   - `SOURCE <filename>`
-  - `COPY <tableName>[(<colums>)] TO <target>[ WITH <options>[ AND <options>...]]`
+  - `COPY <tableName>[(<colums>)] TO|FROM <target>[ WITH <options>[ AND <options>...]]`
 - Add a method `CassandraConnection.setOptionSet(OptionSet)` to programmatically define a custom compliance mode option
   set on a pre-existing connection.